Readonly handling
* Change test for image existance in cleanup.

* Make sure AwsManaged.async_ready always runs find.

+++ tests/conftest.py
@@ -36,6 +36,8 @@ async def cleanup(ainjector):
     try:
         while True:
             aws_image = await ainjector(AwsImage, name='test-ami*')
+            if not aws_image.mob:
+                break
             await aws_image.delete()
     except (LookupError, NotImplementedError):
         pass
